+++ src/main/resources/sqlScripts/tags_create.sql
@@ -1,5 +1,5 @@
 create table if not exists tags(
-    node_id serial not null references nodes (id),
+    node_id bigint not null references nodes (id),
     k varchar(255),
     v varchar(255),
     primary key(node_id, k)
